backend/requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.42.0
gunicorn==25.1.0
yfinance==1.2.0
cachetools==7.0.5
pydantic-settings==2.13.1
